requirements.txt
appdirs==1.4.3
Django==1.10.6
django-bootstrap3==8.2.1
django-braces==1.11.0
django-jet==1.0.5
packaging==16.8
pyparsing==2.2.0
six==1.10.0
